refactored ticket post
