+++ server/Queries/Queries.cs
@@ -123,32 +123,7 @@ public class Queries
             return false;
         }
     }
-
-    public async Task<bool> CreateTicketTask(NewTicketRecord ticketMessages, PasswordHasher<string> hasher)
-    {
-        try
-        {
-            await using var cmd = _db.CreateCommand(
-                "WITH ticketIns AS (INSERT INTO tickets(category_id, subcategory_id, title, user_id, company_id) "
-                    + "values((SELECT id FROM categories WHERE name = $1 AND company_id = $6), (SELECT id FROM subcategories WHERE name = $2 AND main_category_id = (SELECT id FROM categories WHERE name = $1 AND company_id = $6)), $3, (SELECT id FROM users WHERE email = $4 AND company_id = $6), $6) returning id) "
-                    + "INSERT INTO messages(title, message, ticket_id, user_id) "
-                    + "values ($3, $5, (SELECT id FROM ticketIns), (SELECT id FROM users WHERE email = $4 AND company_id = $6))"
-            );
-            cmd.Parameters.AddWithValue(ticketMessages.CategoryName); //$1
-            cmd.Parameters.AddWithValue(ticketMessages.SubcategoryName); //$2
-            cmd.Parameters.AddWithValue(ticketMessages.Title); //$3
-            cmd.Parameters.AddWithValue(ticketMessages.UserEmail); //$4
-            cmd.Parameters.AddWithValue(ticketMessages.Message); //$5
-            cmd.Parameters.AddWithValue(ticketMessages.CompanyFk); //$6
-            await cmd.ExecuteNonQueryAsync();
-            return true;
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine("Error creating ticket" + ex);
-            return false;
-        }
-    }
+    
 
     public async Task<List<TicketRecord>> GetTicketsAll(string email) //email för den som gjort request används för att få vilket företag
     {
