+++ server/Records/Records.cs
@@ -10,7 +10,7 @@ public record MessagesRecord(int MessageId, string Message, int TicketId, string
 
 public record TicketMessagesRecord(TicketRecord TicketRecord, List<MessagesRecord> Messages);
 
-public record NewTicketRecord(int TicketId, string CategoryName, string SubcategoryName, string Title, string UserEmail, int CompanyFk, int MessageId, string Message);
+//public record NewTicketRecord(int TicketId, string CategoryName, string SubcategoryName, string Title, string UserEmail, int CompanyFk, int MessageId, string Message);
 
 public record GetCustomerSupportEmail(string Email);
