+++ server/TicketRoutes.cs
@@ -1,4 +1,5 @@
 ﻿using Npgsql;
+using Microsoft.AspNetCore.Identity;
 using server.Records;
 
 namespace server;
@@ -40,6 +41,40 @@ public static class TicketRoutes
         }
 
         return Results.Ok(tickets);
+    }
+
+    
+    public record NewTicketRecord(int TicketId, string CategoryName, string SubcategoryName, string Title, string UserEmail, int CompanyFk, int MessageId, string Message);
+
+    public static async Task<IResult> PostTickets(PasswordHasher<string> hasher, HttpContext context,
+        NpgsqlDataSource db)
+    {
+        NewTicketRecord ticketMessages = await context.Request.ReadFromJsonAsync<NewTicketRecord>();
+
+        //if (ticketRequest == null)
+        //return Results.BadRequest();
+        try
+        {
+            await using var cmd = db.CreateCommand(
+                "WITH ticketIns AS (INSERT INTO tickets(category_id, subcategory_id, title, user_id, company_id) "
+                + "values((SELECT id FROM categories WHERE name = $1 AND company_id = $6), (SELECT id FROM subcategories WHERE name = $2 AND main_category_id = (SELECT id FROM categories WHERE name = $1 AND company_id = $6)), $3, (SELECT id FROM users WHERE email = $4 AND company_id = $6), $6) returning id) "
+                + "INSERT INTO messages(title, message, ticket_id, user_id) "
+                + "values ($3, $5, (SELECT id FROM ticketIns), (SELECT id FROM users WHERE email = $4 AND company_id = $6))"
+            );
+            cmd.Parameters.AddWithValue(ticketMessages.CategoryName); //$1
+            cmd.Parameters.AddWithValue(ticketMessages.SubcategoryName); //$2
+            cmd.Parameters.AddWithValue(ticketMessages.Title); //$3
+            cmd.Parameters.AddWithValue(ticketMessages.UserEmail); //$4
+            cmd.Parameters.AddWithValue(ticketMessages.Message); //$5
+            cmd.Parameters.AddWithValue(ticketMessages.CompanyFk); //$6
+            await cmd.ExecuteNonQueryAsync();
+            return Results.Ok(true);
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine("Error creating ticket" + ex);
+            return Results.Ok(false);
+        }
     }
     
 }
